gjenbruke komponent for visning av journalposter og pdf
