+++ src/app/components/formPanel.less
@@ -1,7 +1,6 @@
 @import (reference) '@{coreModulePath}nav-frontend-core/less/_variabler';
 
 .sakstype_punch_form {
-    margin-right: 0.5rem;
     flex-basis: 100%;
     overflow-y: auto;
     border-color: #b7b1a9;
