+++ src/app/containers/omsorgspenger/korrigeringAvInntektsmelding/KorrigeringAvInntektsmeldingContainer.tsx
@@ -6,7 +6,7 @@ import { IIdentState } from 'app/models/types/IdentState';
 import { RootStateType } from 'app/state/RootState';
 import { createOMSKorrigering } from 'app/state/actions/OMSPunchFormActions';
 
-import { SplitView } from '../../../components/SplitView';
+import { JournalpostOgPdfVisning } from '../../../components/JournalpostOgPdfVisning';
 import KorrigeringAvInntektsmeldingForm from './KorrigeringAvInntektsmeldingForm';
 
 export interface KorrigeringAvInntektsmeldingContainerProps {
@@ -26,20 +26,19 @@ const KorrigeringAvInntektsmelding: React.FC<KorrigeringAvInntektsmeldingContain
         });
     }, [søkerId, journalpost]);
     const journalposterFraSoknad = soknad?.journalposter || [];
-
+    const journalposter = Array.from(journalposterFraSoknad);
     return (
-        <SplitView soknad={soknad}>
+        <JournalpostOgPdfVisning journalposter={journalposter}>
             <KorrigeringAvInntektsmeldingForm
                 søkerId={søkerId}
                 søknadId={soknad?.soeknadId || ''}
-                journalposter={Array.from(journalposterFraSoknad)}
+                journalposter={journalposter}
             />
-        </SplitView>
+        </JournalpostOgPdfVisning>
     );
 };
 
 const mapStateToProps = (state: RootStateType) => ({
-    punchState: state.PLEIEPENGER_SYKT_BARN.punchState,
     journalpost: state.felles.journalpost,
     identState: state.identState,
     forbidden: state.felles.journalpostForbidden,
