+++ src/app/containers/pleiepenger/Fordeling/fordeling.less
@@ -1,9 +1,6 @@
 @import (reference) '@{coreModulePath}nav-frontend-core/less/_variabler';
 
 .fordeling-container {
-    flex-basis: 100%;
-    display: flex;
-    flex-direction: row;
     height: 100%;
 }
 
@@ -86,8 +83,7 @@
 .journalikkestottetkopi-checkboks {
     display: flex;
     flex-direction: row;
-    align-items: center;
-    justify-items: center;
+    place-items: center center;
 
     .journalikkestottetkopi-checkboks__hjelpetekst {
         margin-left: 0.5rem;
